Refactor methods for setting the transaction to completed along with the delivery
