+++ controllers/delivery-controller-util.js
@@ -2,13 +2,6 @@
  * the delivery controller.
  */
 
-/* The db file and transaction schema are used for retrieving transaction details. */
-const db = require('../models/db.js');
-const Transaction = require('../models/transaction-schema.js');
-
-/* The transaction controller is used to update the transaction status.*/
-const transactionController = require('./transaction-controller.js');
-
 const deliveryControllerUtil = {
 	/**
 	 * Formats the database results for display in the delivery page.
@@ -61,34 +54,6 @@ const deliveryControllerUtil = {
 
 		/* Return the initialized arrays. */
 		return {ids, dates, customers, dropoffs, statuses};
-	},
-
-	/**
-	 * Sets the status of the transaction corresponding to the selected delivery to "Completed".
-	 *
-	 * @param {Express.Request} req  Object that contains information on the HTTP request from the client.
-	 * @param {Express.Response} res  Object that contains information on the HTTP response from the server.
-	 * @param {string} id	String that contains the delivery ID of the selected delivery.
-	 */
-	completeTransaction: function(req, res, id) {
-		/* The transaction ID corresponding to a delivery has a starting digit of 1 and the same
-		 * succeeding digits as the delivery ID.
-		 */
-		const transactionId = parseInt(id) - 10000000;
-
-		/* Get the original status of the corresponding transaction. */
-		query = {id: transactionId};
-		projection = 'id status';
-
-		db.findOne(Transaction, query, projection, function(result) {
-			/* Assign the retrieved transaction details to the body of the request object
-			 * and set the status of the transaction to "Completed."
-			 */
-			req.body.transactionId = result.id;
-			req.body.transactionStatusOld = result.status;
-
-			transactionController.postEditStatusCompleted(req, res);
-		});
 	}
 };
